+++ skglm/gpu/solvers/jax_solver.py
